NEW: LFMN so far final. BUG: Tower ATC refuses airplanes on active runway BUG: Center ATC error in String.format call BUG: Airplane T-O heading set was incorrect ( floor instead of round) UPD: CompanyFleet needs update according to the new library ESystem BUG: StatsPanel typo removed.

+++ JAtcSimLib/src/eng/jAtcSim/lib/traffic/fleets/CompanyFleet.java
@@ -82,7 +82,7 @@ public class CompanyFleet {
   }
 
   private void updateFleetWeightSum() {
-    this.fleetWeightSum = CollectionUtil.sum(this.types, o -> o.weight);
+    this.fleetWeightSum = CollectionUtil.sum(this.types, o -> (double) o.weight);
   }
 
   private void updateCategoryWeightSum() {
